+++ Service/AB.test.logic.js
@@ -1,5 +1,5 @@
-giconst Util = require("./utils");
-const {ExperimentStorage} = require('../db/experiment.storage');
+const Util = require("./utils");
+const {ExperimentStorage} = require('../repositories/experiment.repository');
 
 const ABcheckAttributes = (endUserReq, experiment)=>{
     const {A, B, C} = experiment.variants_ab;
