+++ Service/route.logic.experiment.js
@@ -1,5 +1,5 @@
 const cron = require('node-cron');
-const {ExperimentStorage} = require('../db/experiment.storage')
+const {ExperimentStorage} = require('../repositories/experiment.repository');
 const ffLogic = require("./feature.logic");
 const abLogic = require("./AB.test.logic");
 const Util = require("./utils");
