+++ controller/experiment.controller.js
@@ -1,13 +1,13 @@
 const {json} = require("express");
-const ExperimentRepository = require("../repositories/experiment.repository");
+const {ExperimentRepository} = require("../repositories/experiment.repository");
 const {PropertyNotFound} = require("../errors/NotFound.errors");
 const {ServerUnableError} = require("../errors/internal.errors");
 const { bodyValidator } = require("../validators/body.validator");
 const { BodyNotSent } = require("../errors/BadRequest.errors");
-const experimentRepository = new ExperimentRepository();
+// const experimentRepository = new ExperimentRepository();
 
 const getAllExperiments = async (req, res) => {
-    const result = await experimentRepository.find();
+    const result = await ExperimentRepository.find();
     if (!result) throw new ServerUnableError("getAllExperiments")
     res.status(200).json({result});
 }
@@ -15,7 +15,7 @@ const getAllExperiments = async (req, res) => {
 const getExperimentById = async (req, res) => {
     if (!req.params.experiment_id) throw new PropertyNotFound("experiment_id");
     const experiment_id = req.params.experiment_id;
-    const result = await experimentRepository.retrieve(experiment_id);
+    const result = await ExperimentRepository.retrieve(experiment_id);
     if (!result) throw new ServerUnableError("getExperimentById")
     res.status(200).json({result});
 }
@@ -23,7 +23,7 @@ const getExperimentById = async (req, res) => {
 const getExperimentsByAccountId = async (req, res) => {
     if (!req.params.account_id) throw new PropertyNotFound("account_id");
     const account_id = req.params.account_id;
-    const result = await experimentRepository.findByAttribute("account_id", account_id);
+    const result = await ExperimentRepository.findByAttribute("account_id", account_id);
     if (!result) throw new ServerUnableError("getExperimentsByAccountId")
     res.status(200).json({result});
 }
@@ -31,7 +31,7 @@ const getExperimentsByAccountId = async (req, res) => {
 const getExperimentsAB = async (req, res) => {
     if (!req.params.account_id) throw new PropertyNotFound("account_id");
     const account_id = req.params.account_id;
-    const result = await experimentRepository.findByTwoAttributes("type", "a-b", "account_id", account_id);
+    const result = await ExperimentRepository.findByTwoAttributes("type", "a-b", "account_id", account_id);
     if (!result) throw new ServerUnableError("getExperimentsAB")
     res.status(200).json({result});
 
@@ -40,7 +40,7 @@ const getExperimentsAB = async (req, res) => {
 const getExperimentsFF = async (req, res) => {
     if (!req.params.account_id) throw new PropertyNotFound("account_id");
     const account_id = req.params.account_id;
-    const result = await experimentRepository.findByTwoAttributes("type", "f-f", "account_id", account_id);
+    const result = await ExperimentRepository.findByTwoAttributes("type", "f-f", "account_id", account_id);
     if (!result) throw new ServerUnableError("getExperimentsFF")
     res.status(200).json({result});
 
@@ -50,14 +50,14 @@ const getExperimentsByDate = async (req, res) => {
     if (!req.query.year && req.query.month) throw new PropertyNotFound("year and month");
     const year = req.query.year;
     const month = req.query.month;
-    const result = await experimentRepository.findByDate(year, month);
+    const result = await ExperimentRepository.findByDate(year, month);
     if (!result) throw new ServerUnableError("getExperimentsByDate")
     res.status(200).json({result});
 }
 
 const createExperiments = async (req, res) => {
     if(!bodyValidator(req)) throw new BodyNotSent();
-    const result = await experimentRepository.create(req.body)
+    const result = await ExperimentRepository.create(req.body)
     if (!result) throw new ServerUnableError("createExperiments")
     res.status(200).json({result});
 }
@@ -65,7 +65,7 @@ const createExperiments = async (req, res) => {
 const updateExperimentsByID = async (req, res) => {
     if (!req.params.experiment_id) throw new PropertyNotFound("experiment_id");
     const experimentID = req.params.experiment_id;
-    const result = await experimentRepository.update(experimentID, req.body)
+    const result = await ExperimentRepository.update(experimentID, req.body)
     if (!result) throw new ServerUnableError("updateExperimentsByID")
     res.status(200).json({result});
 }
@@ -73,7 +73,7 @@ const updateExperimentsByID = async (req, res) => {
 const deleteExperimentsByID = async (req, res) => {
     if (!req.params.experiment_id) throw new PropertyNotFound("experiment_id");
     const experimentID = req.params.experiment_id;
-    const result = await experimentRepository.delete(experimentID)
+    const result = await ExperimentRepository.delete(experimentID)
     if (!result) throw new ServerUnableError("deleteExperimentsByID")
     res.status(200).json({result});
 }
